requirements.txt
streamlit
pinecone-client
google-generative-ai
pypdf
python-docx
langchain-text-splitters